+++ test/unit/views/payment-method-view.js
@@ -84,7 +84,7 @@ describe('PaymentMethodView', function () {
       expect(iconContainer.classList.contains('braintree-method__logo@CLASSNAME')).to.be.false;
     });
 
-    it('sets the inner HTML correctly when the paymentMethod is a new card from Apple Pay', function () {
+    it('sets the inner HTML correctly when the paymentMethod is Apple Pay', function () {
       var iconElement, iconContainer, labelElement;
       var paymentMethod = {
         type: 'ApplePayCard',
@@ -109,13 +109,12 @@ describe('PaymentMethodView', function () {
       expect(iconContainer.classList.contains('braintree-method__logo@CLASSNAME')).to.be.false;
     });
 
-    it('sets the label correctly when the paymentMethod is a vaulted card from Apple Pay', function () {
-      var labelElement;
+    it('sets the inner HTML correctly when the paymentMethod is Venmo', function () {
+      var iconElement, iconContainer, labelElement;
       var paymentMethod = {
-        type: 'ApplePayCard',
+        type: 'VenmoAccount',
         details: {
-          cardType: 'Apple Pay - Visa',
-          lastTwo: '92'
+          username: '@name'
         }
       };
 
@@ -123,10 +122,14 @@ describe('PaymentMethodView', function () {
 
       PaymentMethodView.prototype._initialize.call(this.context);
 
+      iconElement = this.context.element.querySelector('.braintree-method__logo use');
+      iconContainer = this.context.element.querySelector('.braintree-method__logo svg');
       labelElement = this.context.element.querySelector('.braintree-method__label');
 
-      expect(labelElement.textContent).to.contain('Apple Pay');
-      expect(labelElement.querySelector('.braintree-method__label--small').textContent).to.equal('');
+      expect(iconElement.getAttribute('xlink:href')).to.equal('#logoVenmo');
+      expect(labelElement.textContent).to.contain('@name');
+      expect(labelElement.querySelector('.braintree-method__label--small').textContent).to.equal('Venmo');
+      expect(iconContainer.classList.contains('braintree-method__logo@CLASSNAME')).to.be.false;
     });
   });